pyproject.toml
[tool.poetry]
name = "order-flow"
version = "0.1.0"
description = "This app can get info about orders in JSON format through API"
authors = ["Iaroslav <GunGalla@gmail.com>"]
readme = "README.md"
packages = [{include = "order_flow"}]
repository = "https://github.com/GunGalla/order-flow-test"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
python-dotenv = "^1.0.0"
django-extensions = "^3.2.1"
dj-database-url = "^1.3.0"
gunicorn = "^20.1.0"
ipython = "^8.12.0"
psycopg2-binary = "^2.9.6"
coverage = "^7.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
